refactor constants & Toast
