+++ src/components/danizt/constants/characters.js
@@ -1,0 +1,4 @@
+export const NUMBERS = '0123456789'
+export const LOWER_CASE = 'abcdefghijklmnopqrstuvwxyz'
+export const UPPER_CASE = 'ABCDEFGHIJKLMNOPQRSTUVWXYZ'
+export const SPECIAL = '!#$%&()*+,-./:;<=>?@[]^_`{|}~'
