+++ src/components/danizt/utils/passwordManager.js
